-ui is generated vaild openapi yaml
