+++ horreum-api/src/main/java/io/hyperfoil/tools/horreum/api/alerting/ChangeDetection.java
@@ -3,6 +3,7 @@ package io.hyperfoil.tools.horreum.api.alerting;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 import jakarta.validation.constraints.NotNull;
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
 public class ChangeDetection {
     @JsonProperty( required = true )
@@ -11,6 +12,7 @@ public class ChangeDetection {
     public String model;
     @NotNull
     @JsonProperty( required = true )
+    @Schema(implementation = String.class)
     public JsonNode config;
 
     public ChangeDetection() {
