+++ horreum-api/src/main/java/io/hyperfoil/tools/horreum/api/data/ExperimentComparison.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 import jakarta.validation.constraints.NotNull;
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
 public class ExperimentComparison {
 
@@ -12,6 +13,7 @@ public class ExperimentComparison {
     public String model;
     @NotNull
     @JsonProperty( required = true )
+    @Schema(implementation = String.class)
     public JsonNode config;
 
     @NotNull
