+++ horreum-api/src/main/java/io/hyperfoil/tools/horreum/api/data/Label.java
@@ -3,6 +3,7 @@ package io.hyperfoil.tools.horreum.api.data;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 import jakarta.validation.constraints.NotNull;
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
 import java.io.Serializable;
 import java.util.Collection;
@@ -38,6 +39,7 @@ public class Label {
     public static class Value implements Serializable {
         public int datasetId;
         public int labelId;
+        @Schema(implementation = String.class)
         public JsonNode value;
 
         public Value() {
