+++ horreum-api/src/main/java/io/hyperfoil/tools/horreum/api/data/ValidationError.java
@@ -3,11 +3,13 @@ package io.hyperfoil.tools.horreum.api.data;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 import jakarta.validation.constraints.NotNull;
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
 public class ValidationError {
     public int schemaId;
     @NotNull
     @JsonProperty(required = true)
+    @Schema(implementation = String.class)
     public JsonNode error;
 
     public ValidationError() {
