--- horreum-api/src/main/java/io/hyperfoil/tools/horreum/api/services/QueryResult.java
@@ -1,5 +1,0 @@
-package io.hyperfoil.tools.horreum.api.services;
-
-public class QueryResult extends SqlService.JsonpathValidation {
-   public String value;
-}
