+++ horreum-ui/src/main/java/io/hyperfoil/tools/horreum/api/alerting/NotificationSettings.java
