+++ horreum-ui/src/main/java/io/hyperfoil/tools/horreum/api/data/Banner.java
