+++ horreum-ui/src/main/java/io/hyperfoil/tools/horreum/api/data/QueryResult.java
@@ -1,0 +1,5 @@
+package io.hyperfoil.tools.horreum.api.data;
+
+public class QueryResult extends JsonpathValidation {
+   public String value;
+}
