+++ horreum-ui/src/main/java/io/hyperfoil/tools/horreum/api/report/ReportComponent.java
