+++ horreum-ui/src/main/java/io/hyperfoil/tools/horreum/api/report/TableReportConfig.java
