+++ horreum-ui/src/main/java/io/hyperfoil/tools/horreum/api/services/LogService.java
