+++ pom.xml
@@ -30,6 +30,7 @@
         <module>infra/horreum-dev-services</module>
         <module>infra/horreum-infra-common</module>
         <module>horreum-api</module>
+        <module>horreum-ui</module>
         <module>horreum-backend</module>
         <module>horreum-client</module>
         <module>horreum-integration-tests</module>
